Improved add page

+++ routes/add.js
@@ -12,20 +12,24 @@ router.get('/', function(req, res) {
 
 /* POST add page. */
 router.post('/', function(req, res) {
-    console.log('fssfds', req.body.other)
-    var robot = new Robot({
-        teamNumber: req.body.teamNumber,
-        abilities: req.body.abilities,
-        other: req.body.other
-    });
-    if (req.files.image) {
-        fs.readFileSync('uploads/' + req.files.image.name)
-        robot.img.data = fs.readFileSync('uploads/' + req.files.image.name);
-        robot.img.contentType = 'image/png';
-        fs.unlinkSync('uploads/' + req.files.image.name)
+    if (req.body.teamNumber) {
+        console.log('fssfds', req.body.other)
+        var robot = new Robot({
+            teamNumber: req.body.teamNumber,
+            abilities: req.body.abilities,
+            other: req.body.other
+        });
+        if (req.files.image) {
+            fs.readFileSync('uploads/' + req.files.image.name)
+            robot.img.data = fs.readFileSync('uploads/' + req.files.image.name);
+            robot.img.contentType = 'image/png';
+            fs.unlinkSync('uploads/' + req.files.image.name)
+        }
+        robot.save();
+        res.send('Done. It might take a while for it to show up in /robots')
+    } else {
+        res.send('Please enter a team number')
     }
-    robot.save();
-    res.send('Done. It might take a while for it to show up in /robots')
 
 });
